+++ resources/views/errors/503.fortune.php
@@ -1,0 +1,8 @@
+<% extends("errors/Error") %>
+
+<% part("errorTitle") %>
+    Down for maintenance
+<% endpart %>
+<% part("errorDescription") %>
+    Please check back later.
+<% endpart %>